Add Unit tests and fixes for oc_base64
This change adds unit test code for oc_base64 code.

This uses some test vectors (input/output) obtained from RFC-4648

While testing error conditions I found two errors that were not
properly handled in the oc_base64_decode.

- The input is not long enough to be a base64 encoded string.
- The padding character `=` in not at the end of the encoded string

Also changes inputs that represent the size of the array to size_t
instead of int.  Since the return value is also used to detect errors
I left it as int. It may be a good idea to change the return type to
ssize_t sometime in the future.

Small modification to the Makefile. It now places the object files
before the library files. If the Linker scans the library files first
it may not grab a dependency need for the object files. This does not
happen if the object files are first. This modification was made
due to an `undefined reference` error when the base64test code was
added.

+++ include/oc_base64.h
@@ -17,10 +17,11 @@
 #ifndef OC_BASE64_H
 #define OC_BASE64_H
 
+#include <stddef.h>
 #include <stdint.h>
 
-int oc_base64_encode(const uint8_t *input, int input_len,
-                     uint8_t *output_buffer, int output_buffer_len);
-int oc_base64_decode(uint8_t *str, int len);
+int oc_base64_encode(const uint8_t *input, size_t input_len,
+                     uint8_t *output_buffer, size_t output_buffer_len);
+int oc_base64_decode(uint8_t *str, size_t len);
 
 #endif /* OC_BASE64_H */
